+++ gawake-cli/services/system/io.github.kelvinnovais.GawakeServer.conf
@@ -12,3 +12,4 @@
   </policy>
 </busconfig>
 
+<!-- /usr/share/dbus-1/system.d/io.github.kelvinnovais.GawakeServer.conf -->
