+++ gawake-cli/src/dbus-server/dbus-server.h
