+++ gawake-cli/src/dbus-server/meson.build
@@ -1,0 +1,1 @@
+run_target('gdbus-codegenerator', command: 'gdbus-codegenerator.sh')
