+++ gawake-cli/src/gawake-cli/meson.build
@@ -1,0 +1,7 @@
+gawake_cli_sources = files(
+	'main.c',
+	'dbus-client.c',
+
+	'../utils/gawake-types.c',
+	'../dbus-server/dbus-server.c'
+)
