+++ gawake-cli/src/gawake-dbus-server/meson.build
@@ -1,0 +1,10 @@
+gawake_dbus_server_sources = files(
+	'main.c',
+	'database-connection.c',
+
+	'../dbus-server/dbus-server.c',
+
+	'../utils/gawake-types.c',
+	'../utils/validate-rtcwake-args.c',
+	'../utils/get-time.c'
+)
