+++ gawake-cli/src/gawaked/meson.build
@@ -1,0 +1,14 @@
+gawaked_sources = files(
+	'main.c',
+	'scheduler.c',
+	'privileges.c',
+	'week-day.c',
+
+	'../utils/get-time.c',
+	'../utils/debugger.c',
+	'../utils/validate-rtcwake-args.c',
+	'../utils/gawake-types.c',
+
+	'../dbus-server/dbus-server.c',
+
+)
