+++ gawake-cli/src/gawaked/privileges.h
