+++ gawake-cli/src/gawaked/week-day.c
