+++ gawake-cli/src/gawaked/week-day.h
