+++ gawake-cli/src/unused/database-checker.c
