+++ gawake-cli/src/utils/colors.h
