+++ gawake-cli/src/utils/validate-rtcwake-args.h
@@ -1,7 +1,7 @@
 #ifndef VALIDATE_RTCWAKE_ARGS_H_
 #define VALIDATE_RTCWAKE_ARGS_H_
 
-#include "gawake-types.h"
+#include "../utils/gawake-types.h"
 
 int validade_rtcwake_args (RtcwakeArgs *rtcwake_args);
